+++ docs/changelog.md
@@ -1,5 +1,14 @@
 # TBD
 
+# 0.11.2
+### Features
+* Added constants which contains the version of this library, which is updated automatically during the release process of this repo
+    * Go: `KurtosisApiVersion`
+    * Typescript: `KURTOSIS_API_VERSION`
+
+### Changes
+* Add explicit Apache-2 licensing
+
 # 0.11.1
 ### Features
 * Ported `services` .go files, including both services and config-factory, to typescript inside `typescript/lib/services` directory